test: Add test for parameters reset
