+++ tests/platform/test_load.py
@@ -1,18 +1,14 @@
 import inspect
 import os
-import shutil
 from pathlib import Path
 
 import pytest
 
-from qibolab._core.dummy.platform import create_dummy_hardware
 from qibolab._core.platform import Hardware, Platform
-from qibolab._core.platform.platform import PARAMETERS
 from qibolab.platform import (
     PLATFORM,
     PLATFORMS_PATH,
     create_platform,
-    initialize_parameters,
     load_hardware,
     locate_platform,
 )
@@ -31,17 +27,6 @@ def test_envvar_unset(monkeypatch):
     monkeypatch.delenv(PLATFORMS_PATH)
     with pytest.raises(RuntimeError, match="unset"):
         _ = locate_platform("ciao")
-
-
-@pytest.fixture
-def dummy_hardware(monkeypatch, tmp_path: Path) -> str:
-    name = "dummy_hardware"
-    parameters = initialize_parameters(hardware=create_dummy_hardware())
-    platform = tmp_path / name
-    shutil.copytree(Path(__file__).parent / name, platform)
-    (platform / PARAMETERS).write_text(parameters.model_dump_json(indent=4))
-    monkeypatch.setenv(PLATFORMS_PATH, str(tmp_path))
-    return name
 
 
 def test_create_platform_from_hardware(dummy_hardware: str):
