+++ tests/platform/test_parameters.py
@@ -1,6 +1,10 @@
+import pytest
+
 from qibolab._core.components.channels import IqChannel
 from qibolab._core.components.configs import IqConfig, IqMixerConfig, OscillatorConfig
-from qibolab._core.platform.load import create_platform
+from qibolab._core.platform.load import create_platform, locate_platform
+from qibolab._core.platform.parameters import reset_parameters
+from qibolab._core.platform.platform import PARAMETERS
 from qibolab._core.pulses.pulse import Pulse, Readout
 from qibolab.platform import Hardware, initialize_parameters
 
@@ -44,3 +48,20 @@ def test_parameters_initialization():
     assert isinstance(parameters.configs["0/drive"], IqConfig)
     assert isinstance(parameters.configs["mixer/ciao"], IqMixerConfig)
     assert isinstance(parameters.configs["lo/come"], OscillatorConfig)
+
+
+def test_parameters_reset(dummy_hardware: str):
+    # ensure no parameters
+    path = locate_platform(dummy_hardware)
+    (path / PARAMETERS).unlink(missing_ok=True)
+
+    with pytest.raises(FileNotFoundError):
+        create_platform(dummy_hardware)
+
+    # reset parameters
+    reset_parameters(dummy_hardware, natives=["RX"])
+
+    # load platform
+    platform = create_platform(dummy_hardware)
+    assert all(q.RX is not None for q in platform.natives.single_qubit.values())
+    assert all(q.MZ is None for q in platform.natives.single_qubit.values())
